Lots of fixes
the error resulting in molecules leaving the simulation space when hitting a cylinder mantle should be removed now
transition from one cylidner into another face to face seems to work quite well
for absorbing scernarios there is still some deviation, reason unknown so far
the rare segmentation faults in the region initialisation should be fixed

+++ src/region.h
@@ -231,7 +231,7 @@ struct region { // Region boundary parameters
 	// this region to allow non-rectangle faces
 	// Only needed if this region is microscopic
 	// TODO added for cylinder integraion, might not be necessary
-	int ** boundRegionFaceShape;
+	short ** boundRegionFaceShape;
 
 	// 2D array of boundary coordinates for faces in boundary regions that overlap
 	// this region.
